app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
argcomplete==3.6.2
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
boto3==1.38.5
botocore==1.38.5
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.39.0
greenlet==3.2.1
griffe==1.7.3
groq==0.23.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-anthropic==0.3.12
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.3.34
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.34
logfire-api==3.14.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
mdurl==0.1.2
mistralai==1.7.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.76.0
opentelemetry-api==1.32.1
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai==0.1.8
pydantic-ai-slim==0.1.8
pydantic-evals==0.1.8
pydantic-graph==0.1.8
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
s3transfer==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
tavily-python==0.7.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
upstash-redis==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
